few changes

+++ src/main/java/in/co/examsadda/crud/repository/OptionRepository.java
@@ -1,5 +1,6 @@
 package in.co.examsadda.crud.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -12,7 +13,7 @@ public interface OptionRepository extends CrudRepository<Option, Integer> {
 
 	public Option findByOptionId(int optionId);
 	
-	public List<Option> findAllByQuestionIdFk(Integer questionId);
+	public List<Option> findAllByQuestionIdFk(BigInteger questionId);
 	
 	public List<Option> findAllBySectionIdFk(int sectionId);
 }
